+++ lib/src/list/type.dart
@@ -7,12 +7,6 @@ typedef AppendCallback = Future<void> Function();
 /// Placeholder state changed callback function, used in `PlaceholderText`
 typedef PlaceholderStateChangedCallback = void Function(PlaceholderState);
 
-/// get page data function, used in `AppendableListView` (appendable)
-typedef GetPageDataFunction<T> = Future<List<T>> Function({int page});
-
-/// get non-page data function, used in `RefreshableListView` and `RefreshableStaggeredGridView` (refreshable)
-typedef GetNonPageDataFunction<T> = Future<List<T>> Function();
-
 /// data model for `AppendableSeriesListView` and `AppendableSeriesStaggeredGridView`
 class SeriesData<T, U> {
   List<T> data;
@@ -20,6 +14,12 @@ class SeriesData<T, U> {
 
   SeriesData({@required this.data, @required this.maxId});
 }
+
+/// get non-page data function, used in `RefreshableListView` and `RefreshableStaggeredGridView` (refreshable)
+typedef GetNonPageDataFunction<T> = Future<List<T>> Function();
+
+/// get page data function, used in `AppendableListView` (appendable)
+typedef GetPageDataFunction<T> = Future<List<T>> Function({int page});
 
 /// get series data function, used in `AppendableListView` (appendable)
 typedef GetSeriesDataFunction<T, U> = Future<SeriesData<T, U>> Function({U maxId});